Fixed LineItems in Transactions

+++ src/Transactions/EpayLineItem.php
@@ -8,15 +8,15 @@ class EpayLineItem extends Epay
 {
     public String $product_key;
     public String $name;
-    public Double $cost;
+    public Float $cost;
     public Int $qty;
     public String $description;
     public String $sku;
     public Bool $taxable;
-    public Double $tax_amount;
+    public Float $tax_amount;
     public String $tax_rate;
     public String $discount_rate;
-    public Double $discount_amount;
+    public Float $discount_amount;
     public String $location_key;
     public String $commodity_code;
 }
